Extract view query component out
